requirements.txt
alembic==1.5.4
apache-airflow==2.0.1
apache-airflow-providers-cncf-kubernetes==1.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-microsoft-mssql==1.0.1
apache-airflow-providers-odbc==1.0.1
apache-airflow-providers-sqlite==1.0.1
apispec==3.3.2
argcomplete==1.12.2
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
cached-property==1.5.2
cachetools==4.2.1
cattrs==1.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.5
defusedxml==0.6.0
dill==0.3.3
dnspython==1.16.0
docutils==0.16
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-auth==1.26.1
graphviz==0.16
gunicorn==19.10.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
jsonschema==3.2.0
kubernetes==11.0.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
natsort==7.1.1
numpy==1.20.1
oauthlib==3.1.0
openapi-spec-validator==0.2.9
openpyxl==3.0.6
pandas==1.2.2
pendulum==2.1.2
prison==0.1.3
psutil==5.8.0
#psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pylint==2.6.2
pymssql==2.1.5
pyodbc==4.0.30
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
pytzdata==2020.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rich==9.2.0
rsa==4.7.1
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0